requirements.txt
tornado==6.1
requests==2.25.1
websocket_client==1.1.0
